requirements.txt
google-generativeai
python-pptx
python-dotenv
requests